requirements.txt
requests
beautifulsoup4
pandas
openai
scikit-learn
flask
langchain
rug
react
axios
python-dotenv